# Conflicts:
#	drivers/flash/flash_stm32_qspi.c
#	drivers/flash/spi_nor.c
#	drivers/rtc/rtc_ll_stm32.c
#	drivers/sdhc/sdhc_spi.c
#	drivers/sensor/sensor_shell.c
#	drivers/spi/spi_ll_stm32.c
#	include/zephyr/drivers/sensor.h
#	include/zephyr/drivers/spi.h
#	include/zephyr/kernel_structs.h
#	include/zephyr/modem/chat.h
#	include/zephyr/net/buf.h
#	include/zephyr/net/coap.h
#	include/zephyr/net/dns_resolve.h
#	include/zephyr/net/ethernet.h
#	include/zephyr/net/igmp.h
#	include/zephyr/net/net_context.h
#	include/zephyr/net/net_event.h
#	include/zephyr/net/net_if.h
#	include/zephyr/net/net_ip.h
#	include/zephyr/net/net_mgmt.h
#	include/zephyr/net/offloaded_netdev.h
#	include/zephyr/net/phy.h
#	include/zephyr/net/ppp.h
#	include/zephyr/net/socketcan.h
#	include/zephyr/net/wifi_mgmt.h
#	include/zephyr/smf.h
#	kernel/init.c
#	lib/smf/smf.c
#	subsys/mgmt/mcumgr/grp/img_mgmt/src/zephyr_img_mgmt.c
#	subsys/modem/modem_cmux.c
#	subsys/modem/modem_ppp.c
#	subsys/net/ip/ipv6.c
#	subsys/net/ip/net_if.c
#	subsys/net/lib/dns/llmnr_responder.c
#	subsys/net/lib/dns/mdns_responder.c
#	subsys/net/lib/dns/resolve.c
#	west.yml

+++ subsys/net/lib/mqtt/mqtt_encoder.c
@@ -94,7 +94,7 @@ static int pack_uint16(uint16_t val, struct buf_ctx *buf)
  */
 static int pack_utf8_str(const struct mqtt_utf8 *str, struct buf_ctx *buf)
 {
-	if ((buf->end - buf->cur) < GET_UT8STR_BUFFER_SIZE(str)) {
+	if ((uintptr_t)(buf->end - buf->cur) < GET_UT8STR_BUFFER_SIZE(str)) {
 		return -ENOMEM;
 	}
 
@@ -102,7 +102,7 @@ static int pack_utf8_str(const struct mqtt_utf8 *str, struct buf_ctx *buf)
 		 (uint32_t)GET_UT8STR_BUFFER_SIZE(str), (void *)buf->cur, (void *)buf->end);
 
 	/* Pack length followed by string. */
-	(void)pack_uint16(str->size, buf);
+	(void) pack_uint16((uint16_t)str->size, buf);
 
 	memcpy(buf->cur, str->utf8, str->size);
 	buf->cur += str->size;
